NP-1858 Move actions to own component
